feat(MCA-12): build-up Episode Page (#21)
Built-up episode page without HTML5 audio-player due to limitations from server side. Spent 4 hours on finding the way how to achieve it, but seems like the only way to do so is build a server and do the magic with cURL (not sure if it is legal to do so).